requirements.txt
python-telegram-bot
trafilatura
google-genai
python-dotenv
